finishe basic ui
