+++ .gitignore
@@ -349,3 +349,4 @@ MigrationBackup/
 # Ionide (cross platform F# VS Code tools) working folder
 .ionide/
 .idea
+/Stat.Itok.Client/Properties/launchSettings.json
