+++ Stat.Itok.Client/Client/Program.cs
@@ -1,7 +1,7 @@
 using Havit.Blazor.Components.Web;
 using Microsoft.AspNetCore.Components.Web;
 using Microsoft.AspNetCore.Components.WebAssembly.Hosting;
-using Stat.Itok.Client;
+using System.Text.Json;
 
 namespace Stat.Itok.Client
 {
@@ -13,8 +13,20 @@ namespace Stat.Itok.Client
             builder.RootComponents.Add<App>("#app");
             builder.RootComponents.Add<HeadOutlet>("head::after");
             builder.Services.AddHxServices();
-
-            builder.Services.AddScoped(sp => new HttpClient { BaseAddress = new Uri(builder.HostEnvironment.BaseAddress) });
+            builder.Services.AddSingleton(_ => new JsonSerializerOptions()
+            {
+                PropertyNameCaseInsensitive = true,
+            });
+            builder.Services.AddHxMessenger();
+            builder.Services.AddHxMessageBoxHost();
+            builder.Services.AddScoped(sp =>
+            {
+                var client = new HttpClient
+                {
+                    BaseAddress = new Uri(builder.Configuration["API_Prefix"] ?? builder.HostEnvironment.BaseAddress)
+                };
+                return client;
+            });
 
             await builder.Build().RunAsync();
         }
