+++ Stat.Itok.Client/Shared/MainLayout.razor
@@ -1,5 +1,7 @@
 ﻿@inherits LayoutComponentBase
 
+<HxMessenger Position="ToastContainerPosition.TopEnd" />
+<HxMessageBoxHost />
 <div class="page">
     <main>
         <article class="content px-4">
