+++ Stat.Itok.Client/Usings.cs
@@ -1,0 +1,1 @@
+global using Stat.Itok.Shared;
