+++ Stat.Itok.Client/_Imports.razor
@@ -10,3 +10,4 @@
 @using Stat.Itok.Client.Shared
 @using Havit.Blazor.Components.Web
 @using Havit.Blazor.Components.Web.Bootstrap
+@using Stat.Itok.Shared
