+++ Stat.Itok.Client/staticwebapp.config.json
@@ -1,0 +1,3 @@
+{
+
+}
