+++ Stat.Itok.Client/wwwroot/appsettings.Development.json
@@ -1,0 +1,3 @@
+{
+  "API_Prefix": "http://localhost:7071"
+}
