+++ Stat.Itok.Core/Entities.cs
@@ -12,45 +12,6 @@ public record IMinkFCalcApiResp
     public string Timestamp { get; set; }
 }
 
-public record NinAccessTokenInfo
-{
-    public string AccessToken { get; set; }
-    public string IdToken { get; set; }
-}
-
-public record NinUserInfo
-{
-    public string Nickname { get; set; }
-    public string Id { get; set; }
-    public string Lang { get; set; }
-    public string Country { get; set; }
-    public string Birthday { get; set; }
-}
-
-public record NinTokenCopyInfo
-{
-    public string TokenCopyUrl { get; set; }
-    public string AuthCodeVerifier { get; set; }
-    public string RedirectUrl { get; set; }
-}
-
-public record NinAuthContext
-{
-    public NinTokenCopyInfo TokenCopyInfo { get; set; }
-    public string SessionToken { get; set; }
-    public string GameToken { get; set; }
-    public string BulletToken { get; set; }
-    public NinUserInfo UserInfo { get; set; }
-    public NinAccessTokenInfo AccessTokenInfo { get; set; }
-    public string PerGameToken { get; set; }
-}
-
-public enum PreCheckResult
-{
-    Ok = 1,
-    AutoRefreshed = 2,
-    NeedBuildFromBegin = 3,
-}
 
 public enum BattleType
 {
@@ -379,43 +340,3 @@ public class StatInkPostBattleSuccess
     public string Url { get; set; }
 }
 
-public class ApiResp
-{
-    public static ApiResp<TData> OkWith<TData>(TData data)
-    {
-        return ApiResp<TData>.OkWith(data);
-    }
-
-    public static ApiResp<TData> Error<TData>(string errorMsg = "failed")
-    {
-        return ApiResp<TData>.Error(errorMsg);
-    }
-}
-
-public class ApiResp<TData> : ApiResp
-{
-    public ApiResp()
-    {
-    }
-
-    public ApiResp(TData data, bool result = true, string msg = "success")
-    {
-        Data = data;
-        Result = result;
-        Msg = msg;
-    }
-
-    public string Msg { get; set; }
-    public bool Result { get; set; }
-    public TData Data { get; set; }
-
-    public static ApiResp<TData> OkWith(TData data)
-    {
-        return new ApiResp<TData>(data);
-    }
-
-    public static ApiResp<TData> Error(string errorMsg = "failed")
-    {
-        return new ApiResp<TData>(default, false, errorMsg);
-    }
-}
