+++ Stat.Itok.Func/Usings.cs
@@ -1,0 +1,1 @@
+global using Stat.Itok.Shared;
