+++ Stat.Itok.Tests/Usings.cs
@@ -1,1 +1,2 @@
 global using Microsoft.VisualStudio.TestTools.UnitTesting;
+global using Stat.Itok.Shared;
